refactor(api): split WorkflowsService into focused workflow services (#754)
WorkflowsService had grown to 2,653 lines with 9 constructor deps across
five unrelated concerns. Split behavior-preserving into:

- WorkflowTemplateSeederService: the 12 idempotent per-org ensure* seeding
  methods, with the 8 near-identical template-loop seeders deduplicated
  into one generic race-safe helper
- LegacyWorkflowStepRunner: the step-based (pre-node) execution engine,
  with the inline topological scheduler decomposed into named methods;
  kept (not deleted) because step-only workflows still reach it via
  WorkflowSchedulerService, migrated cron workflow_execution rows, and
  the manual/webhook compat fallbacks
- WorkflowWebhookService: webhook credential generation/rotation, lookup
  and public trigger path; findByWebhookId now filters config->>'webhookId'
  in SQL instead of loading every workflow and scanning in JS
- WorkflowRunControlService: partial runs, resume-from-failed, credit
  validation, execution logs

WorkflowsService keeps CRUD/templating, lifecycle, node locks and the
ownership guards: 529 lines, 4 injected deps, no method over 80 lines.

Adds partial expression index workflows_config_webhook_id_idx backing the
new webhook lookup (CREATE INDEX CONCURRENTLY, applied by deploy).

generateWebhook now builds the trigger URL from ConfigService.apiUrl
(GENFEEDAI_API_URL) — the same source getWebhookInfo already used —
instead of a divergent raw process.env.API_URL read.

Closes #754
