+++ apps/server/api/scripts/seeds/trend-notification-workflows.seed.ts
@@ -19,7 +19,7 @@ import { resolve } from 'node:path';
 import process from 'node:process';
 import { fileURLToPath } from 'node:url';
 import { AppModule } from '@api/app.module';
-import { WorkflowsService } from '@api/collections/workflows/services/workflows.service';
+import { WorkflowTemplateSeederService } from '@api/collections/workflows/services/workflow-template-seeder.service';
 import { TREND_NOTIFICATION_WORKFLOW_TEMPLATES } from '@api/collections/workflows/templates/trend-notification-workflows.template';
 import { PrismaService } from '@api/shared/modules/prisma/prisma.service';
 import { Logger } from '@nestjs/common';
@@ -83,7 +83,7 @@ async function main(): Promise<void> {
   });
 
   try {
-    const workflowsService = app.get(WorkflowsService);
+    const workflowSeeder = app.get(WorkflowTemplateSeederService);
     const prisma = app.get(PrismaService);
 
     const where: Record<string, unknown> = { isDeleted: false };
@@ -119,7 +119,7 @@ async function main(): Promise<void> {
         continue;
       }
 
-      await workflowsService.ensureTrendNotificationWorkflows(
+      await workflowSeeder.ensureTrendNotificationWorkflows(
         organization.userId,
         organization.id,
       );
