+++ apps/server/api/src/collections/content-schedules/services/content-schedules.service.ts
@@ -207,8 +207,8 @@ export class ContentSchedulesService extends BaseService<
     organizationId: string,
     scheduleId: string,
   ): Promise<void> {
-    const workflowsService = await this.getWorkflowsService();
-    if (!workflowsService) {
+    const workflowSeeder = await this.getWorkflowTemplateSeeder();
+    if (!workflowSeeder) {
       return;
     }
 
@@ -221,7 +221,7 @@ export class ContentSchedulesService extends BaseService<
       return;
     }
 
-    await workflowsService.ensureContentScheduleWorkflow(
+    await workflowSeeder.ensureContentScheduleWorkflow(
       userId,
       organizationId,
       scheduleId,
@@ -232,12 +232,12 @@ export class ContentSchedulesService extends BaseService<
     organizationId: string,
     scheduleId: string,
   ): Promise<void> {
-    const workflowsService = await this.getWorkflowsService();
-    if (!workflowsService) {
+    const workflowSeeder = await this.getWorkflowTemplateSeeder();
+    if (!workflowSeeder) {
       return;
     }
 
-    await workflowsService.disableContentScheduleWorkflow(
+    await workflowSeeder.disableContentScheduleWorkflow(
       organizationId,
       scheduleId,
     );
@@ -254,7 +254,7 @@ export class ContentSchedulesService extends BaseService<
     return organization?.userId ?? undefined;
   }
 
-  private async getWorkflowsService(): Promise<
+  private async getWorkflowTemplateSeeder(): Promise<
     | {
         disableContentScheduleWorkflow: (
           organizationId: string,
@@ -272,9 +272,11 @@ export class ContentSchedulesService extends BaseService<
       return undefined;
     }
 
-    const { WorkflowsService } = await import(
-      '../../workflows/services/workflows.service'
+    const { WorkflowTemplateSeederService } = await import(
+      '../../workflows/services/workflow-template-seeder.service'
     );
-    return this.moduleRef.get(WorkflowsService, { strict: false });
+    return this.moduleRef.get(WorkflowTemplateSeederService, {
+      strict: false,
+    });
   }
 }
