+++ apps/server/api/src/collections/workflows/controllers/workflow-execution.controller.spec.ts
@@ -1,6 +1,7 @@
 import type { AuthenticatedUser as User } from '@api/auth/interfaces/authenticated-user.interface';
 import { WorkflowExecutionController } from '@api/collections/workflows/controllers/workflow-execution.controller';
 import { WorkflowExecutorService } from '@api/collections/workflows/services/workflow-executor.service';
+import { WorkflowRunControlService } from '@api/collections/workflows/services/workflow-run-control.service';
 import { WorkflowSchedulerService } from '@api/collections/workflows/services/workflow-scheduler.service';
 import { WorkflowsService } from '@api/collections/workflows/services/workflows.service';
 import { RolesGuard } from '@api/helpers/guards/roles/roles.guard';
@@ -23,14 +24,17 @@ describe('WorkflowExecutionController', () => {
 
   const mockWorkflowsService = {
     archiveWorkflow: vi.fn(),
-    executePartial: vi.fn(),
     findMutableOwnedOrThrow: vi.fn(),
-    getExecutionLogs: vi.fn(),
     lockNodes: vi.fn(),
     publishWorkflowLifecycle: vi.fn(),
-    resumeFromFailed: vi.fn(),
     setThumbnail: vi.fn(),
     unlockNodes: vi.fn(),
+  };
+
+  const mockWorkflowRunControlService = {
+    executePartial: vi.fn(),
+    getExecutionLogs: vi.fn(),
+    resumeFromFailed: vi.fn(),
     validateCredits: vi.fn(),
   };
 
@@ -54,6 +58,10 @@ describe('WorkflowExecutionController', () => {
       controllers: [WorkflowExecutionController],
       providers: [
         { provide: WorkflowsService, useValue: mockWorkflowsService },
+        {
+          provide: WorkflowRunControlService,
+          useValue: mockWorkflowRunControlService,
+        },
         {
           provide: WorkflowExecutorService,
           useValue: mockWorkflowExecutorService,
