+++ apps/server/api/src/collections/workflows/services/workflow-scheduler.service.ts
@@ -1,7 +1,7 @@
 import { WorkflowExecutionsService } from '@api/collections/workflow-executions/services/workflow-executions.service';
 import type { WorkflowDocument } from '@api/collections/workflows/schemas/workflow.schema';
+import { LegacyWorkflowStepRunner } from '@api/collections/workflows/services/legacy-workflow-step-runner.service';
 import { WorkflowExecutorService } from '@api/collections/workflows/services/workflow-executor.service';
-import { WorkflowsService } from '@api/collections/workflows/services/workflows.service';
 import { ConfigService } from '@api/config/config.service';
 import { PrismaService } from '@api/shared/modules/prisma/prisma.service';
 import { WorkflowExecutionTrigger, WorkflowStatus } from '@genfeedai/enums';
@@ -32,7 +32,7 @@ export class WorkflowSchedulerService implements OnModuleInit {
     private readonly logger: LoggerService,
     private readonly schedulerRegistry: SchedulerRegistry,
     private readonly configService: ConfigService,
-    private readonly workflowsService: WorkflowsService,
+    private readonly legacyWorkflowStepRunner: LegacyWorkflowStepRunner,
     private readonly workflowExecutionsService: WorkflowExecutionsService,
     private readonly workflowExecutorService: WorkflowExecutorService,
   ) {}
@@ -231,7 +231,7 @@ export class WorkflowSchedulerService implements OnModuleInit {
             { triggeredBy: 'schedule' },
             WorkflowExecutionTrigger.SCHEDULED,
           )
-        : this.workflowsService.executeWorkflow(workflowId);
+        : this.legacyWorkflowStepRunner.executeWorkflow(workflowId);
 
       executePromise.catch((error) => {
         this.logger.error(
