+++ apps/server/api/src/seeds/workflow-deployment-backfill.service.ts
@@ -286,37 +286,34 @@ export class WorkflowDeploymentBackfillService {
     userId: string,
     organizationId: string,
   ): Promise<void> {
-    const { WorkflowsService } = await import(
-      '@api/collections/workflows/services/workflows.service'
+    const { WorkflowTemplateSeederService } = await import(
+      '@api/collections/workflows/services/workflow-template-seeder.service'
     );
-    const workflowsService = this.moduleRef.get(WorkflowsService, {
+    const workflowSeeder = this.moduleRef.get(WorkflowTemplateSeederService, {
       strict: false,
     });
 
-    await workflowsService.ensureDailyTrendsDigestWorkflow(
+    await workflowSeeder.ensureDailyTrendsDigestWorkflow(
       userId,
       organizationId,
     );
-    await workflowsService.ensureAdAutomationWorkflows(userId, organizationId);
-    await workflowsService.ensureCampaignOrchestrationWorkflows(
+    await workflowSeeder.ensureAdAutomationWorkflows(userId, organizationId);
+    await workflowSeeder.ensureCampaignOrchestrationWorkflows(
       userId,
       organizationId,
     );
-    await workflowsService.ensureAgentAutopilotWorkflows(
+    await workflowSeeder.ensureAgentAutopilotWorkflows(userId, organizationId);
+    await workflowSeeder.ensureAnalyticsSyncWorkflows(userId, organizationId);
+    await workflowSeeder.ensureContentProductionWorkflows(
       userId,
       organizationId,
     );
-    await workflowsService.ensureAnalyticsSyncWorkflows(userId, organizationId);
-    await workflowsService.ensureContentProductionWorkflows(
+    await workflowSeeder.ensureReplyPollingWorkflows(userId, organizationId);
+    await workflowSeeder.ensureTrendNotificationWorkflows(
       userId,
       organizationId,
     );
-    await workflowsService.ensureReplyPollingWorkflows(userId, organizationId);
-    await workflowsService.ensureTrendNotificationWorkflows(
-      userId,
-      organizationId,
-    );
-    await workflowsService.ensureLivestreamBotWorkflows(userId, organizationId);
-    await workflowsService.ensureSystemActionWorkflows(userId, organizationId);
+    await workflowSeeder.ensureLivestreamBotWorkflows(userId, organizationId);
+    await workflowSeeder.ensureSystemActionWorkflows(userId, organizationId);
   }
 }
